pyproject.toml
[tool.poetry]
name = "penduraai-api"
version = "0.1.0"
description = "API of term sales control app."
authors = ["Hebert <hebertjulio@yahoo.com.br>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.2"
django = "^3.0.6"
djangorestframework-simplejwt = "^4.4.0"
djangorestframework = "^3.11.0"
django-model-utils = "^4.0.0"
djangorestframework-api-key = "^2.0.0"
django-filter = "^2.2.0"
dynaconf = {extras = ["toml"], version = "^2.2.3"}
psycopg2 = "^2.8.5"
whitenoise = "^5.1.0"
redis = "^3.5.2"
channels = "^2.4.0"
channels_redis = "^2.4.2"
daphne = "^2.5.0"
django-silk = "^4.0.1"
dj_database_url = "^0.5.0"
django-redis = "^4.12.1"


[tool.poetry.dev-dependencies]
flake8 = "^3.8.2"
pycodestyle = "^2.6.0"
bandit = "^1.6.2"
faker = "^4.1.0"
ipython = "^7.15.0"
ipdb = "^0.13.2"
poetry = "^1.0.8"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
